// src/Model/productItemsData.model.ts
export interface IProductItemsData {
  salesInvoiceID?: number;
  id: number;
  categoryId?: number;
  name: string;
  description?: string;
  HSN_SAC?: string;
  barCode?: string;
  imagePath?: string;
  ItemImageString?: string;
  ItemImageCss?: string;
  purchasePrice?: number;
  weight?: number;
  weightUnit?: string;
  discount?: number; // Actual discount value; difference in selling price and discounted price
  discountedPrice: number; // For Item History; to be used in transactions
  discountedPriceDuplicate?: number;
  sellingPrice: number;
  sellingPriceDuplicate?: number;
  count: number;
  oldCount?: number;
  isMaintainStock?: boolean;
  isMaintainStockNumber?: number;
  MinStockNotification?: number;
  initialMinStock?: number;
  brand?: string;
  tax?: number;
  viewCount?: boolean;
  addToCart?: boolean;
  mainStock?: boolean;
  currentStock?: number;
  initialStock?: number;
  AllowNegativeStock?: boolean;
  AllowNegativeStockNumber?: number;
  isSellingPriceTaxInclusive?: boolean;
  isSellingPriceTaxInclusiveNumber?: number;
  isPurchasePriceTaxInclusive?: boolean;
  isPurchasePriceTaxInclusiveNumber?: number;
  CreatedDate?: string;
  TaxSlabID1?: number;
  TaxSlabID2?: number;
  TaxSlabID3?: number;
  TaxSlabID4?: number;
  TaxSlabValue1: number;
  TaxSlabValue2: number;
  TaxSlabValue3: number;
  TaxSlabValue4: number;
  ItemMeasurementMasterID?: number;
  ItemMeasurementMasterIDDuplicate?: number;
  InvoiceGenerator?: number;
  TotalQuantity?: number;
  TaxSlabName1?: string;
  TaxSlabName2?: string;
  TaxSlabName3?: string;
  TaxSlabName4?: string;
  PriceUpdate?: boolean;
  TaxSlabData?: number;
  TaxSlabData2?: number;
  TaxSlabPurchase?: number;
  TaxSlabIGSTData?: number;
  TaxSlabCGST?: number;
  TaxSlabSGST?: number;
  TaxSlabIGST?: number;
  TaxSlabCESS?: number;
  TaxSlabCESSData?: number;
  TaxSlabOtherData?: number;
  TaxSlabCESSPercent?: number;
  isActive?: number;
  maximumRetailPrice: number;
  maximumRetailPriceDuplicate?: number;
  SalesInvoiceNo?: number;
  discountType?: string;
  ItemSold?: number;
  customQuantityFlag?: boolean;
  sellingPriceTax?: number;
  taxDuplicate?: number;
  RemainingCount?: number;
  isNegativeStockAllowed?: boolean;
  customQuantityData?: number;
  TaxLabel?: string;
  OtherTaxLabel?: string;
  hsnCode?: string;
}